Moved advisory file processor to csaf package.
